Make ExplosionCrackGeometry
- better than RadialCrackGeometry
- less detritus
