+++ src/game/KesslerSyndrome/elements/Planet.hpp
@@ -81,73 +81,6 @@ namespace game {
     };
 
 
-    SMART_PTR(CrackGeometry);
-
-    class CrackGeometry {
-    public:
-        virtual ~CrackGeometry() {
-        }
-
-        virtual cpBB getBB() const = 0;
-
-        virtual elements::terrain::dpolygon2 getPolygon() const = 0;
-
-        virtual void debugDrawSkeleton() const {
-        }
-
-    protected:
-
-        static elements::terrain::dpolygon2 lineSegmentToPolygon(dvec2 a, dvec2 b, double width);
-
-    };
-
-    class RadialCrackGeometry : public CrackGeometry {
-    public:
-
-        RadialCrackGeometry(dvec2 origin, int numSpokes, int numRings, double radius, double thickness, double variance);
-
-        cpBB getBB() const override {
-            return _bb;
-        }
-
-        elements::terrain::dpolygon2 getPolygon() const override {
-            return _polygon;
-        }
-
-        void debugDrawSkeleton() const override;
-
-    private:
-
-        double _thickness;
-        dvec2 _origin;
-        vector<dvec2> _vertices;
-        vector<vector<pair<size_t, size_t>>> _spokes, _rings;
-        elements::terrain::dpolygon2 _polygon;
-        cpBB _bb;
-
-        elements::terrain::dpolygon2 createPolygon() const;
-
-        elements::terrain::dpolygon2 createPolygon(const pair<std::size_t, std::size_t> &segment) const;
-    };
-
-    class CrackGeometryDrawComponent : public core::DrawComponent {
-    public:
-
-        CrackGeometryDrawComponent(const CrackGeometryRef &crackGeometry);
-
-        ~CrackGeometryDrawComponent() {
-        }
-
-        cpBB getBB() const override;
-
-        void draw(const core::render_state &renderState) override;
-
-    private:
-        CrackGeometryRef _crackGeometry;
-        TriMeshRef _trimesh;
-    };
-
-
 }
 
 #endif /* Planet_hpp */
